feat: merge PR #12 : add low-VRAM mode, fix DINOv3 layer path in pr#12

+++ pixal3d/trainers/flow_matching/mixins/image_conditioned_proj.py
@@ -67,7 +67,7 @@ def project_points_to_image_batch(
     points_homogeneous = torch.cat([points_3d_batch, ones], dim=-1)  # [B, N, 4]
     
     # Compute world to camera transformation matrix
-    world_to_camera = torch.linalg.inv(transform_matrix)  # [B, 4, 4]
+    world_to_camera = torch.linalg.inv(transform_matrix.float()).to(transform_matrix.dtype)  # linalg.inv requires fp32+
     
     # Batch transform to camera coordinate system: [B, N, 4] @ [B, 4, 4]^T -> [B, N, 3]
     points_camera = torch.bmm(points_homogeneous, world_to_camera.transpose(-2, -1))[..., :3]  # [B, N, 3]
